WIP: configurable uid and gid offsets
This patch introduces two new properties: uidoffset and gidoffset.
These properties are used to shift user and group IDs by a configurable
offset. The shift occurs only in memory and what is presented to VFS.
On disk, the IDs are not shifted, which means the offset can be changed
at will without chowning all the files within the dataset.

This feature is useful for unprivileged LXC containers that are using
UID/GID mapping. Create the container _without_ any idmap, then
configure the mapping in LXC and set appropriate uidoffset and gidoffset
on the container's dataset.